+++ LANGSPEC.md
@@ -1,0 +1,114 @@
+## P-Codeの言語仕様
+
+P-Codeはリズムマシンのパターンをテキスト形式で記述するアイデアから発展し、プログラミングの要素を取り入れたライブ・コーディング用の言語である。コードは左から右に解釈され、数値とそれ以外の記号に分けられ、実行される。数値はすべて周波数であり、解釈できない記号はすべてホワイトノイズとして扱われる。
+
+| 記号  | 意味 |
+| ------------- | ------------- |
+| ~ | サイン波  |
+| ^ | 三角波|
+| N | ノコギリ波  |
+| [ | 矩形波 |
+| = | ミュート（休符） |
+| \<CODE\> | CODE部分を繰り返す |
+| NUMBER | 周波数を設定する |
+| +NUMBER | 設定された周波数にNUMBERを足す |
+| -NUMBER | 設定された周波数からNUMBERを引く |
+| *NUMBER | 設定された周波数にNUMBERを掛ける |
+| /NUMBER | 設定された周波数をNUMBERで割る |
+
+## チュートリアル
+
+### 数値と記号
+
+数値は周波数を意味する。記号ひとつが時間の単位(1/30秒)になる。
+
+441Hzのサイン波
+`441~`
+
+11025Hzの矩形波
+`11025[`
+
+数値には小数も使うことができる。
+
+123.456Hzの三角波
+`123.456^`
+
+同じ音を鳴らし続けるには、同じ記号を連続させる。
+
+882Hzのノコギリ波の持続(0.5秒)
+`882NNNNNNNNNNNNNNN`
+
+### 繰り返し
+同じことが繰り返しでも書ける。< > で挟んだ部分を繰り返す。 
+
+< >のネスト(入れ子)で2^(ネストの数)の繰り返し
+`882N`
+`882<N>`
+`882<<N>>`
+`882<<<N>>>`
+`882<<<<N>>>>`
+
+このように、音の長さが 2 倍になっていく。また、カッコが足りない場合は最後に補完される。
+
+`882<<<<N>`
+
+予約されていない文字(記号)以外はすべてホワイトノイズに解釈される。
+
+ノイズの断続(ノイズとミュート(休符)の繰り返し
+`<<<<#=>>>>`
+
+### 四則演算
+
+周波数は加減乗除の演算ができる。
+
+1550Hzの矩形波
+`44100/7/5/2/3*5+1000-500[[[[[`
+
+演算記号を伴わない数値は絶対値とみなされる。
+
+441Hzと882Hzのサイン波の断続の繰り返し 
+`<<<441~~==+441~~==>>>`
+
+上のコードは、以下のようにも書くことができる。
+
+`882<<<-441~~==+441~~==>>>`
+`<<<441~~==*2~~==>>>`
+
+繰り返しを使っていろいろなリズムをつくることができる。
+
+`<<<441~~==*2~~==>>>`
+
+周波数を連続的に変化させることもできる。
+
+サイン波のスイープ(線形)
+`100<<<<<<<<+10~>>>>>>>>`
+
+サイン波の逆スイープ(線形)
+`10000<<<<<<<<-10~>>>>>>>>`
+
+サイン波のスイープ(対数)
+`20<<<<<<*1.1~~~~>>>>>>`
+
+サイン波の逆スイープ(対数)
+`20000<<<<<</1.1~~~~>>>>>>`
+
+### サンプルコード
+
+繰り返しとミュートの組み合わせによるリズムパターン
+`<200N==<<<=50^==><<=800~>2000[>>=*==>>`
+
+数式のように見えるコードもかける
+`<<<100+100=200~>>>`
+
+どんなコードを書いてもエラーにはならない
+`Hello, world!`
+
+沈黙のコード
+`@<<<<<<<<<<=>>>>>>>>>>@`
+
+遭難信号 (SOS) も打てる
+`<<[[==[[==[[==[[[[[[==[[[[[[==[[[[[[==[[==[[==[[========>>`
+
+---
+
+P-Codeはどんなコードを書いてもエラーにはならない。恐れずに（手書きで）コードを書き、僕らの想像を超えるすごい音を聴かせて欲しい。
