+++ README.md
@@ -2,29 +2,28 @@
 
 <a href="https://p-code-magazine.github.io/"><img src="/P-QR-Code.png" width="100"></a>
 
-## [Language specification (in Japanese)](/langspec.md)
+## [Language specification (in Japanese)](/LANGSPEC.md)
 
-## Run prototype
+
+## Build bundle
+
 ```
 npm install
-npm run serve
-```
-Access http://0.0.0.0:8080/
-## Build prototype
-```
 npm run build
 ```
----
-## Run app
+
+## Run example application
+
 Create a Self-Signed SSL Certificate. [How to get https working on your local development environment in 5 minuts](https://www.freecodecamp.org/news/how-to-get-https-working-on-your-local-development-environment-in-5-minutes-7af615770eec/).
+
 ```
-npm install
-npm run serve
+npm run example:serve
 ```
 
 Access https://[LOCAL-SERVER-IP-ADDRESS]:8080/ on your mobile.
 
-## Build app
+## Build example application
+
 ```
-npm run build
+npm run example:build
 ```
