--- app/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "pcode",
-  "version": "1.0.0",
-  "description": "Implementation of P-Code",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "serve": "webpack-dev-server --config webpack.dev.js",
-    "build": "webpack --config webpack.prod.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/p-code-magazine/p-code-magazine.github.io.git"
-  },
-  "author": "p-code-magazine",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/p-code-magazine/p-code-magazine.github.io/issues"
-  },
-  "homepage": "https://github.com/p-code-magazine/p-code-magazine.github.io#readme",
-  "devDependencies": {
-    "https": "^1.0.0",
-    "webpack": "^4.41.2",
-    "webpack-cli": "^3.3.10",
-    "webpack-dev-server": "^3.9.0",
-    "workbox-webpack-plugin": "^4.3.1"
-  },
-  "dependencies": {
-    "p5": "^0.10.2",
-    "tesseract.js": "^2.0.0-beta.2",
-    "tone": "^13.8.25"
-  }
-}
