--- app/webpack.dev.js
@@ -1,27 +1,0 @@
-const path = require('path');
-const fs = require('fs');
-const WorkboxPlugin = require('workbox-webpack-plugin');
-
-module.exports = {
-  mode: 'development',
-  entry: './src/index.js',
-  devServer: {
-    contentBase: './dist',
-    host: '0.0.0.0',
-    hot: true,
-    https: {
-      key: fs.readFileSync('server.key'),
-      cert: fs.readFileSync('server.crt')
-    }
-  },
-  output: {
-    filename: 'main.js',
-    path: path.resolve(__dirname, 'dist'),
-  },
-  plugins: [
-    new WorkboxPlugin.GenerateSW({
-      clientsClaim: true,
-      skipWaiting: true
-    })
-  ]
-};
