+++ example/webpack.prod.js
@@ -1,0 +1,18 @@
+const path = require('path');
+const fs = require('fs');
+const WorkboxPlugin = require('workbox-webpack-plugin');
+
+module.exports = {
+  mode: 'production',
+  entry: './src/index.js',
+  output: {
+    filename: 'main.js',
+    path: path.resolve(__dirname, 'dist'),
+  },
+  plugins: [
+    new WorkboxPlugin.GenerateSW({
+      clientsClaim: true,
+      skipWaiting: true
+    })
+  ]
+};
