+++ package.json
@@ -1,0 +1,41 @@
+{
+  "name": "p-code",
+  "version": "1.0.0",
+  "description": "Implementation of P-Code",
+  "main": "dist/p-code.js",
+  "unpkg": "dist/p-code.js",
+  "module": "lib/index.js",
+  "files": [
+    "lib",
+    "dist",
+    "README.md",
+    "LANGSPEC.md"
+  ],
+  "scripts": {
+    "build": "webpack --env.production",
+    "example:serve": "cd example && webpack-dev-server --config webpack.dev.js",
+    "example:build": "cd example && webpack --config webpack.prod.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/p-code-magazine/p-code.git"
+  },
+  "author": "p-code-magazine",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/p-code-magazine/p-code/issues"
+  },
+  "homepage": "https://github.com/p-code-magazine/p-code#readme",
+  "devDependencies": {
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-server": "^3.9.0",
+    "workbox-webpack-plugin": "^4.3.1",
+    "https": "^1.0.0",
+    "p5": "^0.10.2",
+    "tesseract.js": "^2.0.0-beta.2"
+  },
+  "dependencies": {
+    "tone": "^13.8.25"
+  }
+}
