--- prototype/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "textsp",
-  "version": "1.0.0",
-  "description": "text to sound processing",
-  "private": true,
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "serve": "webpack-dev-server",
-    "build": "webpack"
-  },
-  "author": "hysysk",
-  "license": "ISC",
-  "dependencies": {
-    "p5": "^0.9.0",
-    "tone": "^13.8.25"
-  },
-  "devDependencies": {
-    "webpack": "^4.41.0",
-    "webpack-cli": "^3.3.9",
-    "webpack-dev-server": "^3.8.2"
-  }
-}
