--- prototype/webpack.config.js
@@ -1,15 +1,0 @@
-const path = require('path');
-
-module.exports = {
-  mode: 'development',
-  entry: './src/index.js',
-  devServer: {
-    contentBase: './dist',
-    host: '0.0.0.0',
-    hot: true
-  },
-  output: {
-    filename: 'main.js',
-    path: path.resolve(__dirname, 'dist'),
-  },
-};
